backend/requirements.txt
requests==2.25.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
redis==3.5.3
Flask==1.1.2
Flask-Caching==1.10.1
psycopg2-binary==2.9.3
protobuf==3.20.0
googleapis-common-protos==1.56.0
grpcio==1.48.1
grpcio-tools==1.48.1
environs==9.5.0